// functions/src/types/SIA/enums.ts
export enum eSIALevel {
	PREGRADO = "PREGRADO",
	POSGRADO = "POSGRADO",
}

export enum eSIATypology {
	FUND_OBLIGATORIA = "B FUND. OBLIGATORIA",
	DISC_OBLIGATORIA = "C DISCIPLINAR OBLIGATORIA",
	NIVELACIÓN = "E NIVELACIÓN",
	FUND_OPTATIVA = "O FUND. OPTATIVA",
	TRABAJO_DE_GRADO = "P TRABAJO DE GRADO",
	DISC_OPTATIVA = "T DISCIPLINAR OPTATIVA",
	LIBRE_ELECCIÓN = "L LIBRE ELECCIÓN",
}

/**
 * All UNAL places
 */
export enum eSIAPlace {
	BOGOTÁ = "SEDE BOGOTÁ",
	LA_PAZ = "SEDE DE LA PAZ",
	MEDELLÍN = "SEDE MEDELLÍN",
	MANIZALES = "SEDE MANIZALES",
	PALMIRA = "SEDE PALMIRA",
	TUMACO = "SEDE TUMACO",
	AMAZONÍA = "SEDE AMAZONÍA",
	CARIBE = "SEDE CARIBE",
	ORINOQUÍA = "SEDE ORINOQUÍA",
}

/**
 * Bogota faculties.
 *
 * Bogota is its own faculty for global courses.
 */
export enum eSIABogotaFaculty {
	SEDE_BOGOTÁ = "2000 SEDE BOGOTÁ",
	MEDICINA = "2056 FACULTAD DE MEDICINA",
	ARTES = "2049 FACULTAD DE ARTES",
	IBUN = "2059 INSTITUTO DE BIOTECNOLOGÍA - IBUN",
	ENFERMERÍA = "2054 FACULTAD DE ENFERMERÍA",
	INGENIERÍA = "2055 FACULTAD DE INGENIERÍA",
	ICTA = "2063 INSTITUTO DE CIENCIA Y TECNOLOGÍA DE ALIMENTOS - ICTA",
	ODONTOLOGÍA = "2058 FACULTAD DE ODONTOLOGÍA",
	IEPRI = "2061 INSTITUTO DE ESTUDIOS POLÍTICOS Y RELACIONES INTERNACIONALES - IEPRI",
	IECO = "2064 INSTITUTO DE ESTUDIOS EN COMUNICACIONES - IECO",
	DERECHO = "2053 FACULTAD DE DERECHO, CIENCIAS POLÍTICAS Y SOCIALES",
	INSTITUTO_DE_GENÉTICA = "2062 INSTITUTO DE GENÉTICA",
	AGRONOMÍA = "2048 FACULTAD DE AGRONOMÍA",
	CIENCIAS_ECONÓMICAS = "2051 FACULTAD DE CIENCIAS ECONÓMICAS",
	CIENCIAS = "2050 FACULTAD DE CIENCIAS",
	CIENCIAS_HUMANAS = "2052 FACULTAD DE CIENCIAS HUMANAS",
	CIENCIAS_AGRARIAS = "2728 FACULTAD DE CIENCIAS AGRARIAS",
	MEDICINA_VETERINARIA = "2057 FACULTAD DE MEDICINA VETERINARIA Y DE ZOOTECNIA",
}

/**
 * La Paz faculties.
 *
 * La Paz is its own faculty for global courses.
 */
export enum eSIALaPazFaculty {
	ESCUELA_DE_PREGRADO = "1004 ESCUELA DE PREGRADO DE LA PAZ",
	SEDE_LA_PAZ = "9933 SEDE DE LA PAZ",
}

export type uSIAFaculty = eSIABogotaFaculty | eSIALaPazFaculty;

/**
 * Bogota global (2000) programs
 */
export enum eSIABogotaProgram {
	COMPONENTE_DE_LIBRE_ELECCIÓN = "2CLE COMPONENTE DE LIBRE ELECCIÓN",
	COMPONENTE_ELEGIBLE_POSGRADO = "2CLP COMPONENTE ELEGIBLE POSGRADO",
	CRÉDITOS_PRIMER_PLAN_PARA_DOBLE_TITULACIÓN = "2DTS CRÉDITOS PRIMER PLAN PARA DOBLE TITULACIÓN",
	PROGRAMA_DE_ELECTIVAS = "2PEL PROGRAMA DE ELECTIVAS (L)",
}

/**
 * Science faculty (2050) programs
 */
export enum eSIAScienceProgram {
	BIOLOGÍA = "2513 BIOLOGÍA",
	ESTADÍSTICA = "2514 ESTADÍSTICA",
	FARMACIA = "2515 FARMACIA",
	FÍSICA = "2516 FÍSICA",
	GEOLOGÍA = "2517 GEOLOGÍA",
	MATEMÁTICAS = "2518 MATEMÁTICAS",
	QUÍMICA = "2519 QUÍMICA",
	CC = "2933 CIENCIAS DE LA COMPUTACIÓN",
}

export type uSIAProgram = eSIABogotaProgram | eSIAScienceProgram;